+++ twitter/api.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python
-
 #
-#
-# Copyright 2007 The Python-Twitter Developers
+# Copyright 2007-2016 The Python-Twitter Developers
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -267,6 +265,17 @@ class Api(object):
                                  access_token_key, access_token_secret)
 
         self._config = None
+
+    def GetAccountSettings(self):
+        """Returns dictionary of the current user's account settings.
+
+        Returns:
+            dict: dictionary representing the currently authenticated user's account settings.
+        """
+        url = "%s/account/settings.json" % self.base_url
+        resp = self._RequestUrl(url, 'GET')
+        data = self._ParseAndCheckTwitter(resp.content.decode('utf-8'))
+        return data
 
     def GetHelpConfiguration(self):
         if self._config is None: